fix(models): satisfy ASR evidence lint ratchets
